// google-play-review-analyzer/frontend/src/api/models/fetchAndAnalyzeResponse.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Google Play Review Analyzer
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzedReviewResponse } from './analyzedReviewResponse';

export interface FetchAndAnalyzeResponse {
  app_id: string;
  app_name: string;
  total_analyzed: number;
  reviews: AnalyzedReviewResponse[];
}